Adding wizard for individuals creation needing for new Sample wizard.
In order to describe more accurately observations, users can decide to
import his own individuals set and use the imported one to link with
their observation. The wizard can import from a CSV or excel files.
like in the  CSVLoader, users can choose the targeted column of their
files and create a collection of individuals assigned to a specific
germ plasm.